+++ src/codeblocks/fastimer.ts
@@ -1,8 +1,11 @@
 import { 
-    MarkdownPostProcessorContext,
+    MarkdownPostProcessorContext,    
+    MarkdownRenderer,
 } from 'obsidian';
 
 import DateTime from '../datetime'
+
+import Fastimer from '../main';
 
 import {
     Fast,
@@ -17,7 +20,7 @@ interface FastingZone {
 
 export default class FastimerCodeBlock {
 
-    public static async renderFast(fast: Fast, body: HTMLElement, ctx: MarkdownPostProcessorContext) 
+    public static async renderFast(plugin: Fastimer, fast: Fast, body: HTMLElement, ctx: MarkdownPostProcessorContext) 
     {
         let endTimestamp = fast.currentEndTimestamp > 0 ? fast.currentEndTimestamp : DateTime.now()
         let lines: string[] = []
@@ -26,20 +29,20 @@ export default class FastimerCodeBlock {
 
         if (fast.startTimestamp) {
 
-            lines.push("")
+            lines.push("> ")
 
             this.addFastFrom(lines, fast)
             this.addFastGoal(lines, fast)
 
-            lines.push("")
+            lines.push("> ")
 
             this.addFastingZones(lines, fast, endTimestamp)
 
-            lines.push("")
+            lines.push("> ")
 
             this.addFastProgressBar(lines, fast, endTimestamp)
 
-            lines.push("")
+            lines.push("> ")
 
             this.addFastElapsedTime(lines, fast, endTimestamp)
 
@@ -51,16 +54,16 @@ export default class FastimerCodeBlock {
             }
         }
         
-        body.createEl("pre", {text: lines.join("\n")})
+        MarkdownRenderer.render(plugin.app, lines.join("\n"), body, "", plugin)
     }
 
     private static async addLineWithFastTitle(lines: string[], fast: Fast) {
 
         let text = new Map<FastStatus, string>([
-            [FastStatus.Inactive, "INACTIVE FAST"],
-            [FastStatus.Active, "ACTIVE FAST"],
-            [FastStatus.Completed, "COMPLETED FAST"],
-            [FastStatus.Failed, "FAILED FAST"],
+            [FastStatus.Inactive,  "> [!abstract] Inactive fast"],
+            [FastStatus.Active,    "> [!summary] Active fast"],
+            [FastStatus.Completed, "> [!success] Completed fast"],
+            [FastStatus.Failed,    "> [!failure] Failed fast"],
         ]).get(fast.status);
 
         if (text === undefined) text = "<?>"
@@ -72,14 +75,14 @@ export default class FastimerCodeBlock {
              
         let from = DateTime.timestampToString(fast.startTimestamp)
 
-        lines.push(`From: ${from}`)
+        lines.push(`> **From**: ${from}`)
     }
 
     private static async addFastGoal(lines: string[], fast: Fast) {
 
         let goal = DateTime.timestampToString(fast.plannedEndTimestamp)
 
-        lines.push(`Goal: ${goal}`)
+        lines.push(`> **Goal**: ${goal}`)
     }    
 
     private static addFastingZones(lines: string[], fast: Fast, endTimestamp: number) {
@@ -97,37 +100,37 @@ export default class FastimerCodeBlock {
         let anabolicZone: FastingZone = {
             startTimestamp: anabolicZoneTimestamp,
             endTimestamp: catabolicZoneTimestamp - 1,
-            title: "1. Anabolic    ",
+            title: "1. **Anabolic**",
         }
 
         let catabolicZone: FastingZone = {
             startTimestamp: catabolicZoneTimestamp,
             endTimestamp: fatBurningZoneTimestamp - 1,
-            title: "2. Catabolic   "
+            title: "2. **Catabolic**"
         }
 
         let fatBurningZone: FastingZone = {
             startTimestamp: fatBurningZoneTimestamp,
             endTimestamp: ketosisZoneTimestamp - 1,
-            title: "3. Fat burning "
+            title: "3. **Fat burning**"
         }
 
         let ketosisZone: FastingZone = {
             startTimestamp: ketosisZoneTimestamp,
             endTimestamp: deepKetosisZoneTimestamp - 1,
-            title: "4. Ketosis     "
+            title: "4. **Ketosis**"
         }
 
         let deepKetosisZone: FastingZone = {
             startTimestamp: deepKetosisZoneTimestamp,
             endTimestamp: 0,
-            title: "5. Deep ketosis"
+            title: "5. **Deep ketosis**"
         }
 
         // Rendering:
 
-        lines.push("Fasting zones:")
-        lines.push("")
+        lines.push("> Fasting zones:")
+        lines.push("> ")
 
         this.addFastingZone(lines, fast, anabolicZone, endTimestamp)
         this.addFastingZone(lines, fast, catabolicZone, endTimestamp)
@@ -138,7 +141,7 @@ export default class FastimerCodeBlock {
 
     private static addFastingZone(lines: string[], fast: Fast, zone: FastingZone, endTimestamp: number) {
 
-        let note_text = fast.currentEndTimestamp > 0 ? " ← you were here" : " ← you are here"
+        let note_text = fast.currentEndTimestamp > 0 ? " ==← you were here==" : " ==← you are here=="
         let note = 
             endTimestamp >= zone.startTimestamp 
             && 
@@ -148,7 +151,7 @@ export default class FastimerCodeBlock {
         
         let from = DateTime.timestampToString(zone.startTimestamp)
 
-        lines.push(`${zone.title} ${from}${note}`)
+        lines.push(`> ${zone.title} from ${from}${note}`)
     }
 
     private static async addFastProgressBar(lines: string[], fast: Fast, endTimestamp: number) {
@@ -167,7 +170,7 @@ export default class FastimerCodeBlock {
         let done = "#".repeat(done_len)
         let tail = Math.floor(percent)
     
-        lines.push(`${done}${left} ${tail}%`)
+        lines.push(`> ${done}${left} ${tail}%`)
     }
 
     private static addFastElapsedTime(lines: string[], fast: Fast, endTimestamp: number) {
@@ -177,21 +180,21 @@ export default class FastimerCodeBlock {
          
         let difference = DateTime.timestampsDifference(timestamp1, timestamp2)
 
-        lines.push(`Elapsed time: ${difference}`)
+        lines.push(`> **Elapsed time**: ${difference}`)
     }
 
     private static addFastRemainingTime(lines: string[], fast: Fast, endTimestamp: number) {
          
         let difference = DateTime.timestampsDifference(endTimestamp, fast.plannedEndTimestamp)
 
-        lines.push(`Remaining:    ${difference}`)
+        lines.push(`> **Remaining**: ${difference}`)
     }    
 
     private static addFastExtraTime(lines: string[], fast: Fast, endTimestamp: number) {
          
         let difference = DateTime.timestampsDifference(fast.plannedEndTimestamp, endTimestamp)
 
-        lines.push(`Extra time:   ${difference}`)
+        lines.push(`> **Extra time**: ${difference}`)
     }        
 
 }